+++ docs/conf.py
@@ -39,7 +39,7 @@ extensions = [
     "myst_parser",
 ]
 
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "README.md"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 add_module_names = False
 python_use_unqualified_type_names = True
 toc_object_entries_show_parents = "hide"